refactor: 필드 타입 변경

+++ src/main/java/com/exam/qna/repository/AnswerRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnswerRepository extends JpaRepository<Answer,Integer>, RepositoryUtil {
+public interface AnswerRepository extends JpaRepository<Answer,Long>, RepositoryUtil {
     @Transactional
     @Modifying
     @Query(value = "ALTER TABLE answer AUTO_INCREMENT = 1;", nativeQuery = true)
